+++ src/main/java/socs/network/node/Link.java
@@ -4,14 +4,14 @@ public class Link {
 
   RouterDescription router1;
   RouterDescription router2;
-  int weight=0;
+  short weight=0;
 
   public Link(RouterDescription r1, RouterDescription r2) {
     router1 = r1;
     router2 = r2;
   }
 
-  public Link(RouterDescription r1, RouterDescription r2, int weight) {
+  public Link(RouterDescription r1, RouterDescription r2, short weight) {
     router1 = r1;
     router2 = r2;
     this.weight = weight;
